requirements.txt
django
django-ipware
djangorestframework
markdown
django-filter